add support for saving and restoring state
time for some image-based development!

images are not intended to be compatible between different versions or
build configurations of vulpforth. i tried to prevent accidentally using
obviously-incompatible images by including the length and location of
some of the stuff it copies, but this is not foolproof, use old images
at your own risk.